[refactor] #37: ReportViewModel 을 viewmodel 패키지로 이동
